Plastic Part STLs for the Mendel90 Lasercut
Main changes are:
X-Motor end, the endstop position is raised in order to clear the
pushfit cable connectors and the hole they fit through is enlarged.
Y carriage, raised slightly so if using 3 point bed leveling with a
central hex pillar the screw clears the motor housings
Bearing Holders, sligtly less "grippy" as the zip ties hold them in
anyway.